pyproject.toml
[project]
name = "dataeval-flow"
version = "0.1.0"
description = "DataEval Workflows container for data evaluation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.14"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
dependencies = [
    "dataeval==1.0.4",
    "datasets>=4.0.0",
    "maite-datasets>=0.0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cpu = ["torch>=2.2.0", "torchvision>=0.17.0"]
cu118 = ["torch>=2.2.0", "torchvision>=0.17.0"]
cu124 = ["torch>=2.2.0", "torchvision>=0.17.0"]
cu128 = ["torch>=2.2.0", "torchvision>=0.17.0"]
onnx = ["onnx>=1.15", "onnxruntime>=1.20"]
onnx-gpu = ["onnx>=1.15", "onnxruntime-gpu>=1.20"]
opencv = ["opencv-python-headless>=4.8.0"]
app = ["textual>=3.0"]
# Convenience extras for common combinations
all-cpu = ["torch>=2.2.0", "torchvision>=0.17.0", "onnx>=1.15", "onnxruntime>=1.20", "opencv-python-headless>=4.8.0", "textual>=3.0"]
all-cu118 = ["torch>=2.2.0", "torchvision>=0.17.0", "onnx>=1.15", "onnxruntime-gpu>=1.20", "opencv-python-headless>=4.8.0", "textual>=3.0"]
all-cu124 = ["torch>=2.2.0", "torchvision>=0.17.0", "onnx>=1.15", "onnxruntime-gpu>=1.20", "opencv-python-headless>=4.8.0", "textual>=3.0"]
all-cu128 = ["torch>=2.2.0", "torchvision>=0.17.0", "onnx>=1.15", "onnxruntime-gpu>=1.23.2", "opencv-python-headless>=4.8.0", "textual>=3.0"]

[project.urls]
Repository = "https://gitlab.jatic.net/jatic/aria/dataeval-flow"

[project.scripts]
dataeval-flow = "dataeval_flow.__main__:main"

[dependency-groups]
base = [
    "nox>=2025.5.1",
    "nox-uv>=0.6.2",
    "uv>=0.8.0",
]
lint = [
    { include-group = "base" },
    "ruff>=0.11",
    "codespell[toml]>=2.3",
]
test = [
    { include-group = "base" },
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.1",
    "pytest-xdist>=3.6.1",
    "coverage[toml]>=7.6",
]
type = [
    { include-group = "base" },
    "pyright[nodejs]>=1.1.400",
]
docs = [
    { include-group = "base" },
    "Sphinx>=7.2.6,<9.0.0",
    "sphinx-autoapi>=3.6.0",
    "sphinx-immaterial>=0.12.5",
    "sphinx-design>=0.6.1",
    "sphinx-new-tab-link>=0.8.0",
    "matplotlib>=3.8",
    "myst-nb>=1.2",
    "jupytext>=1.19.1",
    "ipykernel>=7.0",
    "pip>=24.0",
    "dataeval-plots>=0.0.7",
]
docker = [
    { include-group = "base" },
    "jinja2>=3.1",
    "pyyaml>=6.0",
]
dev = [
    { include-group = "lint" },
    { include-group = "type" },
    { include-group = "test" },
]

[tool.uv]
# Pin minimum versions for transitive deps with known CVEs
constraint-dependencies = [
    "pillow>=12.1.1",          # CVE-2026-25990: OOB write via PSD image
    "setuptools>=82.0.0",      # CVE-2026-23949: (jaraco_context) path traversal in tarball()
                               # CVE-2026-24049: (wheel) privilege escalation via unpack
    "onnx>=1.20.1",            # CVE-2026-28500: Untrusted Repository Warnings Suppressed by silent=True
]

conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu118" },
    { extra = "cu124" },
    { extra = "cu128" },
    { extra = "all-cpu" },
    { extra = "all-cu118" },
    { extra = "all-cu124" },
    { extra = "all-cu128" },
  ],
  [
    { extra = "onnx" },
    { extra = "onnx-gpu" },
    { extra = "all-cpu" },
    { extra = "all-cu118" },
    { extra = "all-cu124" },
    { extra = "all-cu128" },
  ]
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cpu", extra = "all-cpu" },
  { index = "pytorch-cu118", extra = "cu118" },
  { index = "pytorch-cu118", extra = "all-cu118" },
  { index = "pytorch-cu124", extra = "cu124" },
  { index = "pytorch-cu124", extra = "all-cu124" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu128", extra = "all-cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cpu", extra = "all-cpu" },
  { index = "pytorch-cu118", extra = "cu118" },
  { index = "pytorch-cu118", extra = "all-cu118" },
  { index = "pytorch-cu124", extra = "cu124" },
  { index = "pytorch-cu124", extra = "all-cu124" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu128", extra = "all-cu128" },
]

# --- Build System ---

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/dataeval_flow"]

[tool.hatch.build.targets.wheel]
include = ["src/dataeval_flow"]

[tool.hatch.build.targets.wheel.sources]
"src/dataeval_flow" = "dataeval_flow"

# --- Pytest ---

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "--pythonwarnings=ignore::DeprecationWarning",
    "--verbose",
    "--durations=20",
    "--durations-min=1.0",
]
markers = [
    "required: marks tests for required functionality [TR-4-H-3]",
    "optional: marks tests for optional functionality [TR-4-H-4]",
]

# --- Coverage ---

[tool.coverage.run]
source = ["src/dataeval_flow"]
branch = true
concurrency = ["multiprocessing"]
parallel = true

[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    ": \\.\\.\\.",
    "if TYPE_CHECKING:",
]
include = ["*/src/dataeval_flow/*"]
fail_under = 90
show_missing = true

[tool.coverage.html]
directory = "output/htmlcov"

[tool.coverage.xml]
output = "output/coverage.xml"

# --- Pyright ---

[tool.pyright]
reportMissingImports = false

# --- Ruff [CR-3-G-2] ---

[tool.ruff]
exclude = [
    ".devcontainer",
    ".github",
    ".vscode",
    ".jupyter_cache",
    "*env*",
    "output",
    "_build",
    ".tox",
    ".nox",
    "prototype",
]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "F", "E", "W", "C90", "I", "N", "D1", "D200", "D201", "D211", "D403", "D419",
    "UP", "YTT", "ANN", "S", "BLE", "B", "A", "C4", "T10", "ISC", "ICN",
    "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID252", "ARG", "FIX", "PD",
    "FLY", "NPY", "RUF100", "PERF",
]
ignore = [
    "ANN401",   # Any type - required for MAITE dataset compatibility
    "C408",
    "C416",
]
fixable = ["ALL"]
per-file-ignores = { "tests/*" = ["ANN", "S101", "D", "SLF001", "ARG002", "PT019"], "docs/*" = ["D", "E402", "E501", "S101"] }

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.isort]
known-first-party = ["dataeval_flow"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- Codespell ---

[tool.codespell]
skip = "./*env*,./output,./htmlcov,./.nox,uv.lock,./docs/source/notebooks/*,./cosign.pub,Dockerfile.*,*.onnx"
ignore-words-list = ["statics"]
